Feature/sprint4-2

+++ kkpedia/src/Components/PersonalPost.js
@@ -1,30 +1,36 @@
 import React, { useState, useEffect } from "react";
 import firebase from "../utils/firebase";
 import styled from "styled-components";
+import Swal from "sweetalert2";
 import RenderPost from "./RenderPost";
-import cancel from "../img/cancel.png";
+import cancel from "../img/trash.png";
+
 const ProfileContainer = styled.div`
-	margin-top: 5vmin;
+	outline: 5px solid black;
 	padding: 4vmin;
 `;
 
 const Container = styled.div`
 	position: relative;
-	/* outline: 1px solid black; */
-	display: flex;
 	flex-direction: column;
+	margin-top: 2vmin;
+	display: flex;
 `;
 
 const DeletePost = styled.div`
 	background-image: url(${cancel});
 	background-repeat: no-repeat;
 	background-size: 100%;
-	width: 3vmin;
-	height: 3vmin;
+	width: 4vmin;
+	height: 4vmin;
 	position: absolute;
-	top: 10px;
-	right: 65px;
+	top: 4vmin;
+	right: 9vmin;
 	cursor: pointer;
+	@media screen and (max-width: 1200px) {
+		right: 7vmin;
+		top: 5vmin;
+	}
 `;
 
 function PersonalPost() {
@@ -50,19 +56,32 @@ function PersonalPost() {
 	}, []);
 
 	const handleDelete = (e) => {
-		db.collection("posts")
-			.doc(`${e.target.dataset.id}`)
-			.delete()
-			.then(() => {
-				alert("被刪除了留言已回不來了😢😢");
-			})
-			.catch((error) => {
-				console.error("Error removing document: ", error);
-			});
+		Swal.fire({
+			title: "確定要刪除嗎?",
+			text: "刪除就回不來了喔!",
+			icon: "warning",
+			showCancelButton: true,
+			confirmButtonColor: "#3085d6",
+			cancelButtonColor: "#d33",
+			confirmButtonText: "Yes!",
+		}).then((result) => {
+			if (result.isConfirmed) {
+				db.collection("posts")
+					.doc(`${e.target.dataset.id}`)
+					.delete()
+					.then(() => {
+						// alert("被刪除了留言已回不來了😢😢");
+						Swal.fire("刪除成功!", "被刪除了留言已回不來了😢😢", "success");
+					})
+					.catch((error) => {
+						console.error("Error removing document: ", error);
+					});
+			}
+		});
 	};
 
 	return (
-		<ProfileContainer>
+		<>
 			{postData.map((item) => {
 				return (
 					<Container key={item.id}>
@@ -71,7 +90,7 @@ function PersonalPost() {
 					</Container>
 				);
 			})}
-		</ProfileContainer>
+		</>
 	);
 }
